% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_shelves.R
\name{expand_shelves}
\alias{expand_shelves}
\title{Long-ify data by bookshelf tags}
\usage{
expand_shelves(data)
}
\arguments{
\item{data}{\verb{[df]} Goodreads data export with "Bookshelves" column.}
}
\value{
data \verb{[df]} Long form of input data with one row per Bookshelf per
book.
}
\description{
List each book once for each bookshelf tag it has.
}
\details{
Tagging books with different bookshelves is a useful way to categorize books
on Goodreads. Books can be on multiple bookshelves. This function creates a
long version of the data where each bookshelf-book combination is its own
row.
}
\examples{
# This example uses our built-in books dataset.
expand_shelves(books)
}
