% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_dates.R
\name{fix_dates}
\alias{fix_dates}
\title{Fix dates from \code{get_read()}}
\usage{
fix_dates(df)
}
\arguments{
\item{df}{the output of \code{get_read()}, a data.frame}
}
\value{
a tibble with the relevant columns formatted as \code{Date} vectors
}
\description{
The output of \code{get_read()} is generally very clean,
however the columns corresponding to dates can be a little messy.
These columns are character vectors and should be in mdy format.
However, it is possible for the date and/or month to be missing.
}
\details{
This function attempts to clean these columns,
making some assumptions:
missing days are set as the first of the month;
missing months are set to January.
}
\examples{
library("magrittr") # for pipe

get_profile_url("bookononism") \%>\%
  get_read() \%>\%
  fix_dates()
}
