% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_read.R
\name{get_read}
\alias{get_read}
\title{Get tidy tibble of a user's reading history}
\usage{
get_read(url)
}
\arguments{
\item{url}{A URL pointing to a user's profile, the output of \code{get_profile_url()}}
}
\value{
a tibble
}
\description{
Get tidy tibble of a user's reading history
}
\examples{
library("magrittr") # for pipe

get_profile_url("bookononism") \%>\%
  get_read()
}
