% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodreadr-package.R
\docType{package}
\name{goodreadr-package}
\alias{goodreadr}
\alias{goodreadr-package}
\title{goodreadr: Plot Your Reading History}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ErinBecker/goodreadr}
  \item \url{https://erinbecker.github.io/goodreadr/}
  \item Report bugs at \url{https://github.com/ErinBecker/goodreadr/issues}
}

}
\author{
\strong{Maintainer}: Erin Becker \email{erinstellabecker@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
