% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_books_api.R
\name{google_books_api}
\alias{google_books_api}
\alias{query_google_books_api}
\alias{google_cover}
\alias{google_isbn}
\title{Querying and parsing data from the Google Books API}
\usage{
query_google_books_api(title, author)

google_cover(data)

google_isbn(data)
}
\arguments{
\item{title}{Book title}

\item{author}{Book author}

\item{data}{parsed JSON data from \code{query_google_books_api()}}
}
\value{
\code{query_google_books_api()} returns a nested list with the results from the API query.
\code{google_cover()} and \code{google_isbn()} return character vectors of length 1 with the relevant information
}
\description{
Various functions for querying and parsing the results of the Google Books API.
\code{query_google_books_api()} takes the title and author of a book as input and returns a nested list
with the results from the API call. The \verb{google_*()} functions parse the nested list, returning
the desired information as character vectors of length 1.
}
\examples{
data <- query_google_books_api("The Obelisk Gate", "Jemisin, N.K.")
google_cover(data)
google_isbn(data)
}
