% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodreadR-package.R
\docType{package}
\name{goodreadR-package}
\alias{goodreadR}
\alias{goodreadR-package}
\title{goodreadR: Scrape and Tidy Goodreads User Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Scrape and tidy user data from goodreads.com with tools from `rvest` and other tidyverse pacakges.
}
\author{
\strong{Maintainer}: James Otto \email{jamesotto852@gmail.com} (\href{https://orcid.org/0000-0002-0665-2515}{ORCID})

}
\keyword{internal}
